better response to errors

+++ site_admin_token.py
@@ -1,5 +1,6 @@
 import authx.auth
 import os
+import sys
 import getpass
 from settings import get_env
 
@@ -34,9 +35,10 @@ def get_site_admin_token(username=None, password=None, refresh_token=None):
 
         if "error" in credentials:
             try:
-                os.remove("tmp/site-admin-refresh-token")
-            except FileNotFoundError:
-                pass
+                if os.path.isfile("tmp/site-admin-refresh-token"):
+                    os.remove("tmp/site-admin-refresh-token")
+                else:
+                    return credentials
             except Exception as e:
                 print(str(e))
                 print(type(e))
@@ -55,4 +57,7 @@ def get_site_admin_token(username=None, password=None, refresh_token=None):
 
 if __name__ == "__main__":
     result = get_site_admin_token()
+    if 'dict' in str(type(result)):
+        print(result)
+        sys.exit(1)
     print(result)
